apps/doc-worker/requirements.txt
celery[rabbitmq]==5.3.4
pydantic==2.5.3
pydantic-settings==2.1.0
orjson==3.11.4
pillow>=10.2.0
numpy>=1.26.4,<3.0.0
opencv-python-headless>=4.9.0.80
pymupdf==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
transformers==4.57.1
onnxruntime==1.17.1
sentencepiece>=0.1.99
asyncpg>=0.29.0,<0.31.0
sqlalchemy==2.0.25
alembic==1.13.1
tenacity>=8.2.3,<10.0.0
python-dotenv==1.0.0
boto3==1.34.0

# Testing dependencies
pytest>=9.0.0
pytest-asyncio>=1.3.0
pytest-cov>=7.0.0
pytest-mock>=3.12.0

